Delegate CKVSI.delete to CellDeleter
